+++ projects/_04_direct_payment_recipients/Dockerfile_and_requirements/Dockerfile
@@ -8,7 +8,6 @@ ENV PYTHONPATH=/app
 
 # Dependencies
 COPY projects/_04_direct_payment_recipients/fargate/utils /app/projects/_04_direct_payment_recipients/fargate/utils
-COPY projects/_04_direct_payment_recipients/direct_payments_column_names.py /app/projects/_04_direct_payment_recipients/direct_payments_column_names.py
 COPY projects/_04_direct_payment_recipients/direct_payments_config_polars.py /app/projects/_04_direct_payment_recipients/direct_payments_config_polars.py
 COPY polars_utils /app/polars_utils
 COPY utils/column_names /app/utils/column_names/
