+++ projects/_04_direct_payment_recipients/direct_payments_configuration.py
@@ -5,7 +5,7 @@ os.environ["SPARK_VERSION"] = "3.5"
 
 from pyspark.sql.types import FloatType, IntegerType, StructField, StructType
 
-from projects._04_direct_payment_recipients.direct_payments_column_names import (
+from utils.column_names.direct_payments_column_names import (
     DirectPaymentColumnNames as DP,
 )
 from utils.column_values.categorical_column_values import ContemporaryCSSR
