+++ projects/_04_direct_payment_recipients/fargate/estimate_direct_payments.py
@@ -1,7 +1,7 @@
 import polars as pl
 
 from polars_utils import utils
-from projects._04_direct_payment_recipients.direct_payments_column_names import (
+from utils.column_names.direct_payments_column_names import (
     DirectPaymentColumnNames as DP,
 )
 from projects._04_direct_payment_recipients.direct_payments_config_polars import (
