+++ projects/_04_direct_payment_recipients/fargate/utils/estimate_direct_payments_utils/create_summary_table.py
@@ -1,6 +1,6 @@
 import polars as pl
 
-from projects._04_direct_payment_recipients.direct_payments_column_names import (
+from utils.column_names.direct_payments_column_names import (
     DirectPaymentColumnNames as DP,
 )
