+++ projects/_04_direct_payment_recipients/fargate/utils/estimate_direct_payments_utils/estimate_service_users_employing_staff.py
@@ -1,7 +1,7 @@
 import polars as pl
 
 from polars_utils.utils import coalesce_with_source_labels
-from projects._04_direct_payment_recipients.direct_payments_column_names import (
+from utils.column_names.direct_payments_column_names import (
     DirectPaymentColumnNames as DP,
 )
 from projects._04_direct_payment_recipients.fargate.utils.estimate_direct_payments_utils.calculate_rolling_mean import (
