+++ projects/_04_direct_payment_recipients/fargate/utils/estimate_direct_payments_utils/merge_cornwall_and_isles_of_scilly.py
@@ -1,6 +1,6 @@
 import polars as pl
 
-from projects._04_direct_payment_recipients.direct_payments_column_names import (
+from utils.column_names.direct_payments_column_names import (
     DirectPaymentColumnNames as DP,
 )
 from utils.column_values.categorical_column_values import ContemporaryCSSR
