+++ projects/_04_direct_payment_recipients/jobs/split_pa_filled_posts_into_icb_areas.py
@@ -6,7 +6,7 @@ from pyspark.sql import DataFrame, Window
 from pyspark.sql import functions as F
 from pyspark.sql.types import StringType
 
-from projects._04_direct_payment_recipients.direct_payments_column_names import (
+from utils.column_names.direct_payments_column_names import (
     DirectPaymentColumnNames as DPColNames,
 )
 from projects._04_direct_payment_recipients.direct_payments_configuration import (
