+++ projects/_04_direct_payment_recipients/tests/fargate/utils/test_calculate_remaining_variables.py
@@ -4,7 +4,7 @@ import polars as pl
 import polars.testing as pl_testing
 
 import projects._04_direct_payment_recipients.fargate.utils.estimate_direct_payments_utils.calculate_remaining_variables as job
-from projects._04_direct_payment_recipients.direct_payments_column_names import (
+from utils.column_names.direct_payments_column_names import (
     DirectPaymentColumnNames as DP,
 )
