+++ projects/_04_direct_payment_recipients/tests/utils/_01_prepare_dpr_utils/test_dpr_prepare_during_year_data.py
@@ -9,7 +9,7 @@ from pyspark.sql.types import (
 )
 
 import projects._04_direct_payment_recipients.utils._01_prepare_dpr_utils.prepare_during_year_data as job
-from projects._04_direct_payment_recipients.direct_payments_column_names import (
+from utils.column_names.direct_payments_column_names import (
     DirectPaymentColumnNames as DP,
 )
 from tests.base_test import SparkBaseTest
