+++ projects/_04_direct_payment_recipients/tests/utils/_01_prepare_dpr_utils/test_dpr_remove_outliers.py
@@ -9,10 +9,10 @@ from pyspark.sql.types import (
 )
 
 import projects._04_direct_payment_recipients.utils._01_prepare_dpr_utils.remove_outliers as job
-from projects._04_direct_payment_recipients.direct_payments_column_names import (
+from utils.column_names.direct_payments_column_names import (
     DirectPaymentColumnNames as DP,
 )
-from projects._04_direct_payment_recipients.direct_payments_column_names import (
+from utils.column_names.direct_payments_column_names import (
     DirectPaymentColumnValues as Values,
 )
 from tests.base_test import SparkBaseTest
