+++ projects/_04_direct_payment_recipients/utils/_01_prepare_dpr_utils/determine_areas_including_carers_on_adass.py
@@ -1,10 +1,10 @@
 from pyspark.sql import DataFrame
 from pyspark.sql import functions as F
 
-from projects._04_direct_payment_recipients.direct_payments_column_names import (
+from utils.column_names.direct_payments_column_names import (
     DirectPaymentColumnNames as DP,
 )
-from projects._04_direct_payment_recipients.direct_payments_column_names import (
+from utils.column_names.direct_payments_column_names import (
     DirectPaymentColumnValues as Values,
 )
 from projects._04_direct_payment_recipients.direct_payments_configuration import (
